# apps/back/docs/api/api.yaml
openapi: 3.1.0
info:
  title: Server API
  version: 0.0.1
  description: Описание интерфейса взаимодействия интерфейса пользователя с сервисом.

  contact:
    name: Фёдор Симоненко
    email: simonenko.f.a@gmail.com

  license:
    url: https://mit-license.org
    name: MIT
servers:
  - url: https://www.adg.dna-technology.ru
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /auth/refresh:
    post:
      tags:
        - Авторизация
      security: []
      operationId: refreshToken
      summary: Обновление токена
      description: Обновляет токен доступа.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./components.yaml#/components/schemas/refresh_payload"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                $ref: "./components.yaml#/components/schemas/tokens"
        "400":
          $ref: "./responses.yaml#/responses/400"
        "404":
          $ref: "./responses.yaml#/responses/404"
        "500":
          $ref: "./responses.yaml#/responses/500"
  /auth/register:
    post:
      tags:
        - Авторизация
      security: []
      operationId: registerUser
      summary: Регистрация пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  description: Электорнная почта
                  example: user@example.com
                password:
                  type: string
                  description: Пароль
                  example: qwerty
      responses:
        "201":
          $ref: "./responses.yaml#/responses/201"
        "400":
          $ref: "./responses.yaml#/responses/400"
        "404":
          $ref: "./responses.yaml#/responses/404"
        "500":
          $ref: "./responses.yaml#/responses/500"
...